+++ src/apps/arrakis/ethereum/arrakis.pool.token-fetcher.ts
@@ -1,4 +1,5 @@
-import { Register } from '~app-toolkit/decorators';
+import { Injectable } from '@nestjs/common';
+
 import { Network } from '~types/network.interface';
 
 import { ARRAKIS_DEFINITION } from '../arrakis.definition';
@@ -8,7 +9,7 @@ const appId = ARRAKIS_DEFINITION.id;
 const groupId = ARRAKIS_DEFINITION.groups.pool.id;
 const network = Network.ETHEREUM_MAINNET;
 
-@Register.TokenPositionFetcher({ appId, groupId, network })
+@Injectable()
 export class EthereumArrakisPoolTokenFetcher extends ArrakisPoolTokenFetcher {
   appId = appId;
   groupId = groupId;