+++ src/apps/rook/rook.module.ts
@@ -3,8 +3,8 @@ import { AbstractApp } from '~app/app.dynamic-module';
 
 import { RookContractFactory } from './contracts';
 import { EthereumRookClaimableContractPositionFetcher } from './ethereum/rook.claimable.contract-position-fetcher';
-import { EthereumRookV2PoolTokenFetcher } from './ethereum/rook.pool-v2.token-fetcher';
-import { EthereumRookV3PoolTokenFetcher } from './ethereum/rook.pool-v3.token-fetcher';
+import { EthereumRookV2PoolTokenFetcher } from './ethereum/rook.v2-pool.token-fetcher';
+import { EthereumRookV3PoolTokenFetcher } from './ethereum/rook.v3-pool.token-fetcher';
 import { EthereumRookXRookTokenFetcher } from './ethereum/rook.x-rook.token-fetcher';
 import { RookAppDefinition, ROOK_DEFINITION } from './rook.definition';