+++ src/apps/tenderize/tenderize.module.ts
@@ -2,11 +2,11 @@ import { Register } from '~app-toolkit/decorators';
 import { AbstractApp } from '~app/app.dynamic-module';
 
 import { ArbitrumTenderizeSwapTokenFetcher } from './arbitrum/tenderize.swap.token-fetcher';
-import { ArbitrumTenderizeTenderTokenFetcher } from './arbitrum/tenderize.tender.tender.token-fetcher';
+import { ArbitrumTenderizeTenderTokenFetcher } from './arbitrum/tenderize.tender.token-fetcher';
 import { TenderizeTokenDefinitionsResolver } from './common/tenderize.token-definition-resolver';
 import { TenderizeContractFactory } from './contracts';
 import { EthereumTenderizeSwapTokenFetcher } from './ethereum/tenderize.swap.token-fetcher';
-import { EthereumTenderizeTenderTokenFetcher } from './ethereum/tenderize.tender.tender.token-fetcher';
+import { EthereumTenderizeTenderTokenFetcher } from './ethereum/tenderize.tender.token-fetcher';
 import { TenderizeAppDefinition, TENDERIZE_DEFINITION } from './tenderize.definition';
 
 @Register.AppModule({